add flex_attention in perf dir
